# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "KAVACH API",
        "description": "Risk, Fraud and Regulatory Intelligence Copilot for Indian Banking",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rules": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "List Rules",
                "description": "List rules with optional filters",
                "operationId": "list_rules_api_rules_get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Page Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rules/{rule_id}/approve": {
            "post": {
                "tags": [
                    "Rules"
                ],
                "summary": "Approve Rule",
                "description": "Approve a rule",
                "operationId": "approve_rule_api_rules__rule_id__approve_post",
                "parameters": [
                    {
                        "name": "rule_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Rule Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApproveRejectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rules/{rule_id}/reject": {
            "post": {
                "tags": [
                    "Rules"
                ],
                "summary": "Reject Rule",
                "description": "Reject a rule",
                "operationId": "reject_rule_api_rules__rule_id__reject_post",
                "parameters": [
                    {
                        "name": "rule_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Rule Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApproveRejectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/rules/{rule_id}": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Get Rule",
                "description": "Get a specific rule",
                "operationId": "get_rule_api_rules__rule_id__get",
                "parameters": [
                    {
                        "name": "rule_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Rule Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rules/upload": {
            "post": {
                "tags": [
                    "Rules"
                ],
                "summary": "Upload Rules",
                "description": "Upload rules from CSV/YAML file",
                "operationId": "upload_rules_api_rules_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_rules_api_rules_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UploadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rules/jobs/{job_id}": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Get Job Status",
                "description": "Get status of a rule upload job",
                "operationId": "get_job_status_api_rules_jobs__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApproveRejectRequest": {
                "properties": {
                    "user": {
                        "type": "string",
                        "title": "User"
                    },
                    "reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reason"
                    }
                },
                "type": "object",
                "required": [
                    "user"
                ],
                "title": "ApproveRejectRequest",
                "description": "Approve/reject request model"
            },
...
            "Body_upload_rules_api_rules_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_rules_api_rules_upload_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "JobStatusResponse": {
                "properties": {
                    "job_id": {
                        "type": "string",
                        "title": "Job Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "progress": {
                        "type": "integer",
                        "title": "Progress"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "job_id",
                    "status",
                    "progress",
                    "message"
                ],
                "title": "JobStatusResponse",
                "description": "Job status response"
            },
...
            "RuleListResponse": {
                "properties": {
                    "rules": {
                        "items": {
                            "$ref": "#/components/schemas/RuleResponse"
                        },
                        "type": "array",
                        "title": "Rules"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "rules",
                    "total",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "RuleListResponse",
                "description": "Rule list response model"
            },
            "RuleResponse": {
                "properties": {
                    "rule_id": {
                        "type": "string",
                        "title": "Rule Id"
                    },
                    "rule_name": {
                        "type": "string",
                        "title": "Rule Name"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    },
                    "typology": {
                        "type": "string",
                        "title": "Typology"
                    },
                    "sql_text": {
                        "type": "string",
                        "title": "Sql Text"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "source_citation": {
                        "type": "string",
                        "title": "Source Citation"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "rule_id",
                    "rule_name",
                    "version",
                    "typology",
                    "sql_text",
                    "status",
                    "source_citation",
                    "created_at"
                ],
                "title": "RuleResponse",
                "description": "Rule response model"
            },
...
            "UploadResponse": {
                "properties": {
                    "job_id": {
                        "type": "string",
                        "title": "Job Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "job_id",
                    "status",
                    "message"
                ],
                "title": "UploadResponse",
                "description": "Upload response"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}